backend/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.2
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.4.1
Django==3.2.7
django-environ==0.6.0
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
